requirements.txt
Django==4.0.5
djangorestframework==3.14.0
django-model-utils==4.2.0 
django-cors-headers==3.13.0 
psycopg2-binary==2.9.3